fix: [BOJ] update 15922.cpp

+++ PS/BOJ/10001-20000/15001-16000/15992.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-int dp[1001][1001];
+long long dp[1001][1001];
 
 int main() {
     int t, n, m;
